+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -19,6 +19,7 @@
 * Removed `Claims::equals` (and `Claims` no longer implements `Comparable`)
 * Removed `Claims::getHash` (and `Claims` no longer implements `Hashable`)
 * Removed `Claims::isEmpty` (you can use `StatementList::isEmpty` instead)
+* Removed `Claims::indexOf` and added `StatementList::getIndexByGuid`
 
 ## Version 2.6.0 (2015-03-08)